# api/docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      message:
        type: string
    type: object
  models.FoodReq:
    properties:
      description:
        type: string
      food_name:
        type: string
      total_capacity:
        type: integer
      union:
        example: piece
        type: string
    type: object
  models.FoodRes:
    properties:
      description:
        type: string
      food_name:
        type: string
      id:
        type: string
      total_capacity:
        type: integer
      union:
        type: string
    type: object
...
  models.ListFoodsRes:
    properties:
      foods:
        items:
          $ref: '#/definitions/models.FoodRes'
        type: array
    type: object
  models.ListProductsRes:
    properties:
      products:
        items:
          $ref: '#/definitions/models.ProductRes'
        type: array
    type: object
  models.ProductReq:
    properties:
      description:
        type: string
      product_name:
        type: string
      total_capacity:
        type: integer
      union:
        type: string
    type: object
  models.ProductRes:
    properties:
      description:
        type: string
      id:
        type: string
      product_name:
        type: string
      total_capacity:
        type: integer
      union:
        type: string
    type: object
  models.Result:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
  description: API for Farmer
  title: Welcome To Farmish API
paths:
...
  /v1/foods:
    get:
      consumes:
      - application/json
      description: Api for List food by page limit and extra values
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: union
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListFoodsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: LIST FOOD
      tags:
      - FOOD
    post:
      consumes:
      - application/json
      description: Api for Create new food
      parameters:
      - description: createModel
        in: body
        name: Food
        required: true
        schema:
          $ref: '#/definitions/models.FoodReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FoodRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: CREATE FOOD
      tags:
      - FOOD
    put:
      consumes:
      - application/json
      description: Api for Update food by food id
      parameters:
      - description: createModel
        in: body
        name: Food
        required: true
        schema:
          $ref: '#/definitions/models.FoodRes'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FoodRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: UPDATE FOOD
      tags:
      - FOOD
  /v1/foods/{id}:
    delete:
      consumes:
      - application/json
      description: Api for Delete food by food ID
      parameters:
      - description: Food ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: DELETE FOOD
      tags:
      - FOOD
    get:
      consumes:
      - application/json
      description: Api for Get food by ID
      parameters:
      - description: Food ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FoodRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: GET FOOD BY FOOD ID
      tags:
      - FOOD
  /v1/products:
    get:
      consumes:
      - application/json
      description: Api for List Product by page limit and extra values
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: union
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListProductsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: LIST PRODUCT
      tags:
      - PRODUCT
    post:
      consumes:
      - application/json
      description: Api for Create new product
      parameters:
      - description: createModel
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.ProductReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: CREATE PRODUCT
      tags:
      - PRODUCT
    put:
      consumes:
      - application/json
      description: Api for Update product by product id
      parameters:
      - description: createModel
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.ProductRes'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: UPDATE PRODUCT
      tags:
      - PRODUCT
  /v1/products/{id}:
    delete:
      consumes:
      - application/json
      description: Api for Delete product by product ID
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: DELETE PRODUCT
      tags:
      - PRODUCT
    get:
      consumes:
      - application/json
      description: Api for Get product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: GET PRODUCT BY PRODUCT ID
      tags:
      - PRODUCT
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"